% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanz_subgrp.R
\name{bzGetSubgrp}
\alias{bzGetSubgrp}
\title{Get subgroup treatment effect estimation and variance}
\usage{
bzGetSubgrp(data.all, var.ey, var.variance, var.cov)
}
\arguments{
\item{data.all}{subject level dataset}

\item{var.ey}{column name in \code{data.all} for estimated treatment effect}

\item{var.variance}{column name in \code{data.all} for variance of subgroup
treatment assignment}

\item{var.cov}{array of column names in \code{dat.all} that corresponds to
binary or ordinal baseline covaraites}
}
\value{
A dataframe with treatment effect estimation and variance for each subgroup
}
\description{
Compute subgroup treatment effect estimation and variance for subgroup effect
summary data. The estimation and variance are combined if there are multiple
record of the same subgroup, defined by the covariates, in the data.
}
