% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanz_present.R
\name{bzRptTbl}
\alias{bzRptTbl}
\title{Summary table of treatment effects}
\usage{
bzRptTbl(lst.stan.rst, dat.sub, var.cov, cut = 0, digits = 3)
}
\arguments{
\item{lst.stan.rst}{list of class \code{beanz.stan} results from
\code{\link{bzCallStan}} for different models}

\item{dat.sub}{dataset with subgroup treatment effect summary data}

\item{var.cov}{array of column names in dat.sub that corresponds to binary or
ordinal baseline covariates}

\item{cut}{cut point to compute the probabiliby that the posterior subgroup
treatment effects is below}

\item{digits}{number of digits in the summary result table}
}
\value{
A dataframe with summary statistics of the model selected by DIC
}
\description{
Compare the DIC from different models and report the summary of treatment effects
based on the model with the smallest DIC value
}
