% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beanz_present.R
\name{r.plot.trace}
\alias{r.plot.trace}
\title{Trace plot of \code{rstan} samples}
\usage{
r.plot.trace(stan.rst, n.eachrow = 2, width = 4, height = 0.8)
}
\arguments{
\item{stan.rst}{a class \code{beanz.stan} object generated by
\code{\link{call.stan}}}

\item{n.eachrow}{number of trace plot each row}

\item{width}{plot size: width}

\item{height}{plot size: height}
}
\description{
Trace plot of \code{rstan} samples for checking the convergence of the MCMC chains
}
\seealso{
\code{\link{call.stan}}
}

